Move get_intent_entity_score into stt/metrics.py
Keep all metric roots (WER, CER, LLM-judge, intent/entity) in
stt/metrics.py. The per-row judge call, prompt, model, and result schema
stay in stt/intent_entity.py as supporting code.
